requirements.txt
requests
xlwt
xmltodict
MyQR
bs4
Pillow